Add methods to retrieve build status (e.g. from Jenkins)
This info is stored under the Stash Build Integration Plugin. Build statuses are associated with a git commit SHA, so can be retrieved by the method `buildStatus(commit)`. A helper method to retrieve the build status for the head of a pull request is also provided, `pullRequestBuildStatus(project, repo, pr_id)`, which uses the pull-request API to find the commit sha, then looks up the build status of that.